backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20

# Google Gemini
google-genai==1.5.0

# MongoDB
motor==3.7.0
pymongo==4.11.3

# Redis
redis[hiredis]==5.2.1

# File processing
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
Pillow==11.1.0

# Monitoring
prometheus-client==0.21.1

# Utilities
python-dotenv==1.0.1
